requirements.txt
paho-mqtt==1.3.0
setproctitle==1.1.10
Jinja2==2.9.6
python-etcd==0.4.5
-e git+https://github.com/TuxEatPi/python-aio-etcd.git#egg=aio_etcd




# dev
#mock==2.0.0
#Sphinx==1.4.8
#sphinx-rtd-theme==0.1.9
#pycodestyle==2.0.0
#pylint==1.6.4
